avoid unnecessary import mindformers
