+++ mindocr/models/backbones/__init__.py
@@ -4,15 +4,6 @@ from ._registry import *
 
 # helpers
 from .builder import *
-from .cls_mobilenet_v3 import *
-from .det_mobilenet import *
-from .det_resnet import *
-from .rec_abinet_backbone import *
-from .rec_master import *
-from .rec_resnet import *
-from .rec_resnet45 import *
-from .rec_svtr import *
-from .rec_vgg import *
 
 __all__ = []
 __all__.extend(builder.__all__)
